Order by person__id when it's going to be used in distinct.
A bad query is somehow getting through the Django ORM to postgres.
The first occurrence of this is on 2016-03-05, which is a couple
of days after we upgraded to Django 1.8, so I imagine it's
something that's changed in the ORM.

Postgres is complaining that a field which is has a distinct on it
needs to be the first in the order by, so I've added an order by
for this field.