Fix for SSL errors on requests to Google's geocoder
We needed this fix for YourNextRepresentative, and it explains transient
errors I was seeing when requesting results from Google's geocoder.  The
background for this, reproduced from the Pombola commit is as follows.

The error is:

  SSLError: [Errno bad handshake] [('SSL routines', 'SSL3_GET_SERVER_CERTIFICATE', 'certificate verify failed')]

... from requests.  I believe that this error is due to the servers we
currently run having OpenSSL 1.0.1 and using a recent version of the
certifi Python package (which contains root certificates).

A full explanation can be found here:

  https://github.com/certifi/python-certifi/issues/26

(found via: http://stackoverflow.com/a/34665344/223092 )

Since upgrading OpenSSL to 1.0.2 will be a pain, pinning certifi to the
old version is the easiest way to fix this error for the moment.
Hopefully we will upgrade the distributions on our servers in the near
future to one that supplies OpenSSL 1.0.2 and then we can switch to a
more recent certificate bundle.