Override InfoBlogView to add old comments for ZA.
There used to be special case code to display the old Disqus comments
for ZA in the info sub-app of pombola. Now that info is being spun out
into its own package this code clearly can't live there any more.

Instead, we subclass the InfoPage view in the ZA views and add
CommentArchiveMixin to it there.