Cope with the ward councillor API being down
Previously, if the Code4SA ward councillor API was down, users would
see an error page, even though the page could still be providing MP and
MPL results.

This is a quick "Friday afternoon" fix to catch such errors and display
an error explaining that the ward councillor results will be missing.