ZA: Quick fix for rep-locator specific code breaking person page tabs
jQuery UI tabs are used on person pages as well as the rep locator
results page - unfortunately some of the code in tabs.js is specific to
the rep locator page, and was causing a Javascript error on person pages
that would stop the JavaScript tab switching from working.

This is a simple fix for that - the code specific to the rep locator is
only run if an element with the expect class of the rep locator tabs
list exists on the page.

Fixes #2156