Hugely simplify JavaScript loading
We found we were wasting huge amounts of time fixing bugs in JavaScript
due to load order bugs, and several developers and designers had
commented on how awkward it was to add simple JavaScript enhancements to
the site.  This is partly because adding JavaScript was complicated by
the following things:

 * Using yepnope (now deprecated) and Modernizr.load to conditionally
   load different JavaScript for desktop and mobile.

 * Having a special addExtraJS method and django-pipeline JavaScript
   tag rendering template to allow django-pipeline-generated JavaScript
   files to be loaded by Modernizr.load.

 * JavaScript sometimes being loaded outside those mechanisms, by
   including the pipeline javascript tag in the page (and in one of
   several places in the page, e.g. <head>, mid-way through <body>,
   end of <body>, etc.

 * Some JavaScript functions were pushed onto
   pombola_run_when_document_ready_array, which was run when the
   document is ready by yepnope.

To make this easier, I've chosen to follow the simple advice here:

  http://www.html5rocks.com/en/tutorials/speed/script-loading/

(in the "Enough! How should I load scripts?" section) to add script
tags at the end of the body element.  Code which needs to be run on
document ready can still be done with jQuery's $(function () { ... });
or $(document).ready(function () { ... }), but since the code now
runs at the end of <body> this will be unnecessary in many cases.
Pages that were trying to call functions to add map markers, for
example, now just output JavaScript to build up the data, which is
then rendered when the map-rendering JavaScript loads.

This means that we can get rid of yepnope and the addExtraJS hacks
and everything to do with JavaScript gets a lot simpler.

We also believe that the mobile-specific JavaScript is no longer
worthwhile, so the same JavaScript is loaded on desktop and mobile.
(There's not much of it anyway, so this should make a negilible
difference.)

Fixes #1686
Fixes #2167