Update Modernizr to a minimal v3.3.1 build, but with SVG detection
Modernizr is now only used for the HTML5shiv, but included lots more
that now isn't needed (e.g. yepnope) and was on an old version anyway.
So, it makes sense to update the version of Modernizr and create a
new minimal build. (The custom Modernizr build in the current version
outputs a URL to reproduce the build, so it'll be easier to update
now as well.

There was some of our own Javascript used for SVG feature detection,
and since this can be done by Modernizr, I've used that instead.

Fixes #1774

Incidentally, there is a strong argument that we shouldn't put
Modernizr in the <head>:

  https://github.com/Modernizr/Modernizr/issues/878#issuecomment-41448059

... and inline the HTML5 Shiv (including html5shiv-printshiv.js).

I'll create an issue about considering that.