Add a parliamentary session filter to position views
Currently all the /position views only return current positions.
However, it's reasonable (and useful) to want to be able to find
positions from earlier parliamentary terms.

This commit adds a ?session=parliamentary-session-slug filter parameter
to the position views; if that is supplied, then only positions whose
dates overlap [1] those of the parliamentary session are returned.

[1] Or possibly overlap, in the cases of positions with approximate
start or end dates.