Allow association of a ParliamentarySession to a position title
This is done by making 'house' (an association with Organisation)
optional and adding an optional 'position_title' attribution (a
PositionTitle) as well.  This is needed to support the National Assembly
in Kenya; how we model this is somewhat controversial, but the current
situation is members are represented by:

  (in 2007-2013)
    position title: Member of Parliament (slug: mp)
    organisation (house): Parliament (slug: parliament)
      ... or in some cases:
    organisation (house): REPUBLIC OF KENYA (slug: republic-of-kenya)

  (in 2013-)
    position title: Member of the National Assembly> (slug: member-national-assembly)
    organisation (house): Parliament (slug: parliament)
      ... or in some cases:
    organisation (house): REPUBLIC OF KENYA (slug: republic-of-kenya)
      ... or in one case:
    organisation (house): National Assembly (slug: national-assembly)

The senate in Kenya, meanwhile, is:

  (in 2013-)
    position title: Senator (slug:senator)
    organisation (house): REPUBLIC OF KENYA (slug: republic-of-kenya)
      ... or in quite a lot of cases:
    organisation (house): Parliament (slug: parliament)

This means that in Kenya, for the moment, finding all people who were
representatives in the senate is best done by associating that session
just with the position title "senator" and not the house.  This commit
makes the 'house' attribute optional and sets the senate session's
house to None.

Related bugs:

  https://github.com/mysociety/pombola/issues/1935
  https://github.com/mysociety/pombola/issues/1967
  https://github.com/mysociety/pombola/issues/2064
  https://github.com/mysociety/pombola/issues/2065