Add a command to mirror ward boundaries from the code4sa MapIt
This command fetches all boundaries of the WD (Ward) type from Code4SA's
MapIt instance and imports them into PA's (unadvertised) MapIt
installation.  As with the other recent changes, this is to improve
lookup times and make the site robust to network failures that might
cause these lookups to fail.