WIP: Add support for mirroring ward councillor data from nearby.code4sa.org
This is another change to reduce the fragility of the site due to its
dependence on remote APIs.  In this case the ward councillor data will
be fetched for every ward in the local MapIt.  n.b. this means that the
south_africa_mirror_mapit_wards command must have been run before
running this script.

This is marked as Work-In-Progress because this API is currently down,
and so I haven't had a chance to test it yet.