Links generated to a sitting should actually show that sitting
Sitting.get_absolute_url would omit the start_time suffix if
start_time was None, but the SittingView would interpret that
URL as referring arbitrarily to the first sitting in that venue
on that date.  This meant that appearances might link to the
wrong sitting, confusingly.

This commit changes that logic so that if the start_time is
missing, it looks for a sitting with start_time None.

Fixes #2195