Use new hover-dropdown component for main menus
The showing/hiding of dropdowns in the main nav is now handled
by .js-hover-dropdown, which gave us an opportunity to refactor
and simplify the CSS styling for the #main-nav across all the
cobrands.

This also includes the following JavaScript enhancements to improve
the experience for touchscreen users and people navigating the page
with a keyboard.

* Highlight active dropdown trigger even without using hover/focus.
* Prevent top level links navigating to new pages on first touch.
* Clicking/pressing outside a dropdown will close the dropdown.
* Allow keyboard users to tab through dropdowns.