Require a known working version of lxml
This should have been in requirements.txt anyway, since it's directly
used by the as_clean_html template filter. (This didn't cause an error,
because it was being installed, as it's a requirement of other
dependent packages, e.g. za-hansard, but it's better to specify the
version we know works - some of the za-hansard tests fail with 3.6.3
and 3.6.4.)