ZA: Show 'must read' posts in sidebar
The posts must be tagged with a tag whose slug is 'must-read'
in order to appear there.

This change updates the django-info-pages package so that we
can use 'posts_by_tag'.

We also added a new template tag to the core template tags
which allows us to return the value corresponding to a key when it
contains characters that are not allowed in template variable
names like ',', '.', '-', etc. since it is very common to have dashes in
tag slugs.