Add the organization kind object to the context of a position page
As per issue 2221, PMG requested to delete the word "any"
but only for the MP Profiles page header, which at the moment
reads "Member of any Parliament".

In this commit, the organization kind object is added to the context,
(similar to the position title), so that we can discriminate between
different position pages.