ZA: Customize the page title on the MP Profiles page
As per issue 2221, PMG requested to delete the word "any"
but only for the MP Profiles page header, which at the moment
reads "Member of any Parliament".

In order to remove "any" from the MP Profiles page header only, we
are adding a conditional that checks for the value of the relevant slugs
and overwrites the page title only for the special case of an MP
Profile page.

Tests check for the presence of both a position title of "Member" and
an organization kind of "Parliament" in order to target that particular
page and not affect any of the other position pages.