ZA: Customize the sub-title on the MP Profiles page
As per issue 2221, PMG requested to delete the subtitle that
reads "Current Position Holders", but only for the MP Profiles page.

In order to remove this subheader from the MP Profiles page only,
we are adding a conditional that checks for the value of the relevant
slugs and shows the subheader only if the page is not an MP
Profile page.

Tests check for the presence of both a position title of "Member" and
an organization kind of "Parliament" in order to target that particular
page and not affect any of the other position pages.