[KE] Add SMS shortcode to Mzalendo homepage
Voter registration is over in Kenya, so the Elections section wording
needed a small tweak, and the IEBC office locator needed removing.

In its place, we add a "Contact us offline" button that displays the
Mzalendo SMS shortcode, and points to an `sms:` URL that will open the
messaging app on most devices that support text messaging.

Fixes #2263.