Always output Facebook app_id meta tag
Some templates, such as blog posts and person pages, override the
`open_graph` template block, which means the `fb:app_id` meta tag isn't
in the HTML for those pages.

To avoid that problem this moves the `fb:app_id` meta tag outside any
template blocks so that it always appears in the HTML output.