ZA: Add a data migration to add ParliamentarySession objects
For Kenya, we've created instances of a `ParliamentarySession` model
to represent a term of a particular house of parliament. (The name is
a bit misleading - it should probably be `Term` instead.) This was
introduced primarily so that we could distinguish places from
different terms with different boundaries. Later we added links to
views of positions of a particular parliament: the position view, when
provided with a `session` query parameter, only includes positions
whose start and end date overlap with the start and end date of the
corresponding `ParliamentarySession`.

In Pombola, this mechanism provides the only easy way to see all the
people who were ever a representative in a particular house for a
given session, which means that since there are no such
`ParliamentarySession`s for South Africa, the EveryPolitician scraper
can only ever find the current representatives.

Essentially the only `/position/` URL that's used on People's Assembly
is the "MP Profiles" page which has been overridden and doesn't
include the session switch link, so adding these ParliamentarySession
objects will only affect pages that aren't linked to by default, but
the following links will still be helpful to the EP scrapers:

* /position/delegate/parliament/ncop/?session=ncop25
* /position/delegate/parliament/ncop/?session=ncop26
* /position/member/parliament/national-assembly/?session=na25
* /position/member/parliament/national-assembly/?session=na26

Note that because these views just look for overlapping dates with a
session, just creating the `ParliamentarySession` objects is enough -
they don't need to be linked with positions by foreign keys. (This is
probably a misfeature.)

PMG have been careful about setting precise start and end dates, so
this should accurately identify who was really a represenative during
each session.