Add a command to add IEBC codes to MapIt counties and constituencies
The recently published candidate lists include the IEBC codes for
counties and constituencies, which are also used in their shapefiles, so
seem to be stable identifiers. We should be able to get these codes from
MapIt lookups, so this command can be used to add those codes to the
MapIt County (DIS) and Constituency (CON) areas.