The Google Feed API is deprecated and no longer works; replace its use
The Google Feed API, which we used to fetch an RSS feed in Javascript
and display it on the homepage, was deprecated and stopped working at
the end of 2016.

There are similar generic services we could use, but it's not clear how
stable or trustworthy those would be, and it's almost as simple to
write our own Javascript for fetching the feed, parsing it and
rendering it as a list, as this commit does.

For this to work there need to be appropriate CORS headers returned by
the server returning the RSS feed. For example, in the case of
Mzalendo, this means that we've added:

  Access-Control-Allow-Origin: http://info.mzalendo.com

... to reponses from http://www.mzalendo.com/blog/

Fixes #2292