Add WriteInPublic client
This is a very simple wrapper around the WriteInPublic API. Currently it
only implements the features that we need for integrating it into
Pombola.

I've added some tests for the basic functionality, but in the interests
of time I've skipped writing tests for the .people() and .answers()
methods, so will need to revisit that at some point in the future.