Add WriteInPublic message creation wizard
This is a multi-step form wizard that walks a user through the steps of
creating a new message to a representative.

It uses django-formtools to roughly replicate the process of writing to
a representative as it currently exists on writeinpublic.com. Once the
user has written a message it uses the WriteInPublic API to send the
message and then takes the user to a confirmation page.