WriteInPublic: Redirect to pending message page at end of wizard
We've enabled moderation on the WriteInPublic instance, which means that
the message isn't available via the API immediately. Rather than
redirecting to a message page instead we redirect to a pending message
page which lets the user know that their message needs to be approved.

This isn't the prettiest page at the moment, but it will at least stop
the immediate errors.