[WriteInPublic] Make view more generic
Rather than hardcoding the Person class all over the place this instead
uses an adapter approach, which means we can pass in different adapters
to the view which can use different models to get the required
information.

This will allow us to have two different WriteInPublic instances, one
that uses `Person` models to write to people and another, which is yet
to be added, which will allow you to write to a committee, which is
represented by the `Organisation` model.