[WriteInPublic] Display the correct app name on the message page
Rather than hard-coding "Write to my MP" on the message page we instead
want to display the name of the current app. To do this I've passed the
app name in as a kwarg to the view, and then pulled that into the
context in the relevant view class and used that in the template.