[ZA] Add page with a list of NA committees
This page lists all NA committees that can be written to, along with
a link to the page showing messages sent to a specific committee.

I originally wanted to show the messages inline on this page, but
that got complicated because that involved making (possibly multiple)
requests to the WriteIt API to gather up all the messages and then link
them to committees. So just linking to the messages page for now seems
like an OK start.