[ZA] Single line of homepage actions on wide screens
Given #2370, it’s pretty obvious we can’t make the buttons any smaller.
But we also can’t make the page any wider. The only thing that can give
is the width of the `<h2>` – either by shortening the text or by
wrapping it onto two lines.

We couldn’t think of a way to sensibly shorten the text, so I’ve gone
with wrapping it onto two lines.

This lets us fit the buttons (at the exact same size they were before)
and the heading (again, at the same font-size) onto a single line on
wide screens.

The CSS is complicated slightly by making sure the layout still looks
good in browsers without flex support. But I think it’s worth it.

Fixes #2393.