Add a URL for an empty page to reuse as a template for the blog
The Mzalendo blog is Wordpress-based, served from our orgsites
repository, and it's always been a problem keeping the header and footer
in that repository (which are largely copy-and-pasted from the Pombola
site, and reuse its CSS) in sync with updates made to the Pombola site
(i.e. in this repository).

To make it easier to generate the Wordpress header and footer from the
current site, this commit adds an endpoint which produces an empty page
we can easily rewrite.