Switch Disqus JS to use the same URL for all pages relating to a Person
Previously there were separate threads on the sub-pages (e.g.
/person/uhuru-kenyatta/contact_details/) from the main page for a person
(e.g. /person/uhuru-kenyatta/). The same applied for places,
organizations, and the peron-specific hansard pages.

This updates the Disqus JavaScript to use the disqus_canonical_url and
disqus_identifier variables we're now adding to the context, which
should be the same for all pages relating to a single person / place /
organization.

It also switches to the currently recommended way of setting the Disqus
JavaScript configuration variables, which is by defining a disqus_config
function which sets this.page.url, this.page.identifier and
this.page.title.