Update the command for generating a Disqus old->new thread URL mapping
This didn't handle all the different cases of URL structures that there
are currently threads for. Also, it suffixed the URLs to go to the
[...]/comments sub-page, but I think it's better now to make the
canonical URL for a thread the base URL of the person/place/organisation,
since on Mzalendo we don't even link to the /comments sub-page any more,
so very few of the existing comments are on /comments threads.

There are still some threads associated with more exotic URLs that we
can't really map - these are typically from caches where you can't infer
the real page from the URL.