[KE] Remove feedback_report_pending from crontab
Looking in the admin _all_ of the feedback since 25 Jun 2014 is still
pending, no one has ever changed the state of any feedback in the admin
(though they might well have read it, which is good).

So I'm going to remove this from the crontab to get one less email each
day to the cron-mzalendo email alias.