Allow places to have images associated with them
This adds the relevant bits to the Place model so that it can have
images associated with it. It works in the same way as images for people
and organisations, you can add images in the admin and then access them
in the template with `.primary_image()`.

It seems that the template for the places list already handles
displaying images for places, including generating an appropriately
sized thumbnail.