[KE] Lookup Hansard speakers by constituency
If a person isn't found using the normal lookup methods then try to look
them up using the new "Constituency, Party initials" format.

At the moment this only uses the constituency, since we know which house
the sitting is for and when it took place, so we can get the correct
constituency back from that, and since they're single member
constituencies they should just have one member associated with them.

In it's current state this still has a few issues, such as not handling
more than one place being returned, and it doesn't pay any attention to
the party initials, but from my testing it seems to pick up most of the
basic cases correctly.