[KE] Make sure Women's Representatives are matched correctly
Because of the way things are modelled the Women's Representatives are
associated with counties from the Senate. This should probably be fixed
at some point, but for now we can just work around it by looking for an
appropriate place and position directly.

This should allow all the simple cases of Women's Representatives using
Constituency and Party initials to be matched in Hansard.