[KE] Hansard: Only use parts of speaker name up to first party initial
Previously we were taking any parts of the name that didn't look like a
party initial, but this was causing an issue where a stray `-` in the
name would end up as part of the place name. So to avoid this only use
the parts up to the position of the first party initial.