[ZA] Fix multiple active tabs on rep locator
The Ward Councillor tab is the default view when you search for an
address, but for some reason the MPs tab was also being marked as
active, which meant that two tabs were receiving the active class when
the rep locator loaded.

This change removes the HTML class for active tabs and leaves it to the
JS to control it.