[ZA] Move position name to the defaults argument
We've already created positions, but some of them have had their names
changed. This is because the ordinal creation has a bug which creates
the 11th as 11st. The name was corrected to "11th Candidate" but the
slug remains "11st_candidate". So just check for positions with the
correct slug, and if they need creating then give them a (possibly
incorrect) name.