Make location of the various artefact and media directories configurable
Rather than having a hardcoded assumption for the location for things
like the virtualenv, gems, media files and caches, make this possible
to configure. The default becomes `data/` in the root of the project
rather than `..` (which assumes that the code has been cloned into an
empty directory).

This is intended to make working with the repository more intuitive.
A developer should be able to clone the repository and run `vargant up`
without having to take further steps.

The name of the python variable containing this has changed from `
root_dir` to `data_dir`. It can be overriden in the applicaton config
by setting `DATA_DIR` in `general.yml`.