Ensure virtualenv location is derived from config
Now that the location of the virtualenv is configurable, we need to
dervive this from config. This adds steps to ensure this is done in
a couple of helper scripts.

Currently this includes a dependency on commonlib and on having
PyYAML available system-wide (this is already listed as a dependency).