Update package dependency file
This removes some really old packages (PostgreSQL 8.4, OpenJDK JRE 6) and
adds some updated versions for Debian Stretch.

It also comments out the PostgreSQL server packages with a note stating
that they are only needed if running the database server locally. This
won't be the case on any of our usual deployments so this prevents
alarming error messages at deploy.