Update paths to Commonlib in the crontab template
As we'll no longer be including commonlib as a submodule, this updates
the PATH set in the crontab template to use our system-wide version and
includes some comments better explaining this dependency.

We may subsequently remove this file and replace it with an example
template as it is specific to mySociety's deployment environment.