Remove CONTRIBUTING.md
This document is very out of date. The bulk of it is to do with
waffle.io, which no longer exists. Other parts of it are about issue
management processes that are no longer relevant.

Going to remove this. If we want to add it back in the future then it's
probably better to start from scratch anyway.