Add writeinpublic_identifier method to Person
We're no longer using EveryPolitician, but for legacy reasons we're
still using EveryPolitician UUIDs in WriteInPublic. We want to continue
using WriteInPublic, but don't have EP identifiers for new people.

Adding this method gives us a standard identifier for each person. It
returns the EP UUID if they have one, otherwise falls back to their ID
in the database.