Add za_hansard app from external repo
This code is only used by People's Assembly, so for now roll it all into
the same codebase so it's easier to develop alongside pombola.

The original repo is hosted at https://github.com/mysociety/za-hansard

You can look at that repository for the history of various files beyond
this commit.