Upgrade sorl-thumbnail and Pillow
When we upgraded Pillow before we started seeing errors like the
following:

     IOError: cannot write mode RGBA as JPEG

As a quick workaround we just downgraded Pillow, but that created a
different error:

     IOError: encoder error -2 when writing image file

I've dug into this a bit more and it seems that the first error is fixed
in sorl-thumbnail 12.4.1. I think I tried upgrading to the latest sorl,
but that isn't compatible with Django 1.8, which is the version we're
currently on. So this change upgrades to the latest version of Pillow,
and to the latest compatible version of sorl.